完成编辑账本页面和接口的开发，并完成对接

+++ app/src/main/java/pers/xyj/accountkeeper/repository/entity/BookVo.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 import java.util.Date
 
 data class BookVo(
-    var id: Long,
+    var id: Int,
 
     var name: String,
